requirements.txt
pyautogui~=0.9.54
pyperclip~=1.11.0
loguru~=0.7.3
python-dotenv~=1.2.1
